+++ services/table-geocoder/lib/internal-geocoder/abstract_query_generator.rb
@@ -1,7 +1,7 @@
 # encoding: utf-8
 
 require_relative 'input_type_resolver'
-require_relative '../../../importer/lib/importer/cartodb_id_query_batcher'
+require_relative '../../../importer/lib/importer/uery_batcher'
 
 module CartoDB
   module InternalGeocoder
